+++ app-kts/src/main/java/com/rossyn/blocktiles/game2048/presentation/activities/GameActivity.kt
@@ -3,8 +3,10 @@ package com.rossyn.blocktiles.game2048.presentation.activities
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class GameActivity: BaseActivity(){
-    override fun onBackPressedCustom() {
+class GameActivity : BaseActivity() {
 
+
+    override fun onBackPressedCustom() {
+        finish()
     }
 }