+++ app-kts/src/main/java/com/rossyn/blocktiles/game2048/presentation/dialogs/scoreBoardDialog.kt
@@ -1,0 +1,4 @@
+package com.rossyn.blocktiles.game2048.presentation.dialogs
+
+class scoreBoardDialog {
+}